Refactor locale handling by introducing useLocale hook and updating components to use it
