+++ next.config.js
@@ -2,8 +2,4 @@ module.exports = {
   images: {
     domains: ['iiif.imaginerio.org', 'imaginerio-images.s3.us-east-1.amazonaws.com'],
   },
-  i18n: {
-    locales: ['en', 'pt'],
-    defaultLocale: 'en',
-  },
 };
